[9.0][FIX]stock: missing ondelete cascade on picking id for pack operations
The current behavior: When trying to delete a cancelled picking, a
message error is desplayed saying that a wrong value is set on pack
operations.
The issue: The `picking_id` field is set as required on the
`stock.pack.operation` without specifying the ondelete parameter. Hence
by default when the picking is deleted the `picking_id` value becomes
null and a DB foreign key constraint is raised.
The solution: add ondelete="CASCADE" constraint on the `picking_id`
field so when a picking in deleted all its pack operations are removed
as well.