[FIX] product: Product Supplier Info wrong in Purchase Order
Steps to reproduce the bug:

- Please use a runbot with 2 companies (with the Admin user):
- Create a new product
- Into this product, create a first line of Vendor and fill it it
(Vendor = China Export and Vendor product code = "Vendor ref Company 1" and put the Company 1 for the Company field)
- Add a second Vendor line into this product
(Keep the same vendor; code = "Vendor ref Company 2" and put the Company 2 for the company field)
- Save your product
- With your demo user logged into the Company 1
- Create a new Purchase Order (vendor = China Export)
- For the product of the first PO line, write the name of your product. You should have 2 results.
- Select the one with displayed name contains "Company 1" (should be the first line)
- When it's selected, the Description (name field) of the purchase order line is always the last
(so in this example, the description contains the product code of the vendor of the other company:
"Vendor ref Company 2" instead of "Vendor ref Company 1").

The function name_get defined on product.product cannot distinguish
which supplier info to take for the display_name because no info is
available for that. So if several Vendor Product Codes are given for
the same company and the same vendor by design it won't work.

opw:803596